requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
pandas>=1.5.0
openpyxl>=3.0.0
flask>=2.0.0
flask-cors>=3.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
schedule>=1.2.0

